requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-talisman==1.1.0
bcrypt==4.0.1
cryptography==41.0.7
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.0.0
sqlalchemy==2.0.21
markupsafe==2.1.3
flask-wtf==1.2.1